bugfix/workaround: argument list too long (E2BIG)
Background
I came across an issue when mass importing files into a sumo arena
The error was:
  mount: Argument list too long

The error didn't have anything to do with sumo, nor aufs, it was the
kernel's E2BIG error code, returned by mount when execve() was not able
to fit the supplied argument list and environment into its buffer.

The problem manifested itself when useraufs-remount was called and the
del/append entries were added to /etc/mtab.

Solution 1:
Set the aufs module parameter brs to 1 and enable CONFIG_AUFS_SYSAUFS.
This would cause aufs to not show the br: (or dirs=) mount entries in
/proc/mounts or /etc/mtab.

Solution 2: (this commit)
Pass the -n option to mount when calling useraufs-remount.
The -n option mounts as normal, without writing to /etc/mtab.